requirements.txt
fastapi>=0.110
uvicorn>=0.23
httpx>=0.25
python-dotenv>=1.0
langchain>=0.3
langgraph>=0.2
langchain-openai>=0.2
langchain-core>=0.3
pydantic>=2.7,<3
